requirements.txt
# General utilitiesFlask==3.1.0python-dotenv==1.0.1requests==2.32.3# OpenAI, Claude AI, and Google AI APIsopenai==1.58.1anthropic==0.42.0google-generativeai==0.8.3# Text-to-Speech and Speech Recognitionazure-cognitiveservices-speech==1.41.1SpeechRecognition==3.13.0pydub==0.25.1# Object Detection and Computer Visionopencv-python==4.10.0.84ultralytics==8.3.56# System and Activity Monitoringpsutil==5.9.5# Twitch Chat Integrationtwitchio==2.10.0# Image Processing and OCRPillow==9.5.0# Stable Diffusion API Integrationtorchvision==0.20.1transformers==4.30.2